added code to fetch nearby fuel stations using geolocation, & rendering fuel prices
