+++ package.json
@@ -19,10 +19,11 @@
   },
   "homepage": "https://github.com/nikoauer/Fuel-Wise#readme",
   "dependencies": {
+    "axios": "^1.5.0",
     "bcrypt": "^5.0.0",
     "connect-session-sequelize": "^7.0.4",
     "dotenv": "^8.2.0",
-    "express": "^4.17.1",
+    "express": "^4.18.2",
     "express-handlebars": "^5.2.0",
     "express-session": "^1.17.1",
     "mysql2": "^2.2.5",
