+++ server.js
@@ -1,7 +1,6 @@
 const express = require('express');
 const path = require('path');
 const session = require('express-session');
-const hbs = require('express-handlebars');
 const authRoutes = require('./routes/authRoutes');
 const fuelRoutes = require('./routes/fuelRoutes');
 const app = express();
@@ -12,7 +11,6 @@ app.use(express.static(path.join(__dirname, 'public')));
 app.use(session({ secret: 'your-secret-key', resave: false, saveUninitialized: true }));
 
 // Set up view engine (Handlebars)
-app.engine('handlebars', hbs.engine);
 app.set('view engine', 'handlebars');
 app.set('views', path.join(__dirname, 'views'));
