+++ views/partials/dashboard.handlebars
@@ -1,4 +1,4 @@
-{{!-- views/dashboard.handlebars --}}
+{{!-- code creates a dynamic list of fuel stations with their names, addresses, prices per unit, and distances from the user's location. --}}
 <h1>Fuel Prices Near You</h1>
 <ul id="fuel-prices">
   {{#each fuelStations}}
