+++ views/partials/fuelPrices.handlebars
@@ -1,3 +1,4 @@
+{{!-- Template generates an HTML table structure for displaying fuel prices, including the fuel station name, location, fuel type, and price per liter/gallon. The values are injected into the template using Handlebars expressions. --}}
 <!DOCTYPE html>
 <html lang="en">
 <head>
